requirements.txt
streamlit==1.36.0
python-dotenv==1.0.1
langchain==0.2.7
langchain-google-genai==1.0.7
neo4j==5.22.0
google-generativeai==0.7.2